// example3/single-spa-vite/vendor-account-ui/src/app/store/pricing/pricing.types.ts
import { IReportsFilesData } from '@middleware/reports/reports.types';
import { IDefaultReportsState, ReportsFilter, SortOrderType } from '@store/root.types';

export const PricingComputedPropertyState = {
  PRICE_LOAD: 'priceLoad',
  PRICELIST_LOAD: 'pricelistLoad',
  DISCOUNTS_LOAD: 'discountsLoad',
} as const;

export interface ISetReportsDataActionState {
  records: number;
  data: IReportsFilesData[];
  computedProperty: keyof typeof PricingComputedPropertyState;
}

export interface ISetReportsPageActionState {
  page: number;
  rows: number;
  computedProperty: keyof typeof PricingComputedPropertyState;
}

export interface ISetReportsSortActionState {
  sortStatus: SortOrderType;
  sortDate: SortOrderType;
  sidx: string;
  computedProperty: keyof typeof PricingComputedPropertyState;
}

export interface IResetReportsFilterState {
  computedProperty: keyof typeof PricingComputedPropertyState;
  key: keyof typeof ReportsFilter;
}

export interface ISetReportsFiltersState {
  filters: Partial<IDefaultReportsState['filters']>;
  computedProperty: keyof typeof PricingComputedPropertyState;
}